remove seek and rewrite operations in state save logic
Prior to this commit, the state save logic operated by first writing a header
and segment descriptions, and then later returning to the start of the output
file and overwriting these with their updated equivalents. By tracking an
expected offset into the output file, this commit omits the first write of these
structures and instead writes out the state in a single pass when all
information is known.

While this commit removes the first redundant write, its main motivation is
actually to remove the seek operation that was required. Seeking the file
descriptor prevents us using certain compression approaches on the saved state.

Note that this commit is intended to have no user-visible effect and the
resulting saved state output should be byte-for-byte identical to its previous
form.